Serial communication library
adjusted the return on read as well as made the code self-finding serial port. and fixed the error with writing lists. Added clear function (will clear the buffer for reading this should be used if the data is off sync suggested to use at the beginning of the code so that there is a clear connection.