Added a button to trigger the pairing process
The server starts in normal mode, where it does not accept incoming pair messages. Pressing the "Pair" button puts the server into pairing mode. Then, pressing the "Pair" button on the button will attempt to pair with the server.
Note that there is a bug where if the server receives a "Disconnect" message while not in pairing mode, it stops working.