Small updates to sender.ino and receiver.ino
receiver.ino: LED is now used to show the button's status. It is turned
on whenever the button is connected to the server and off otherwise. In
the future, it'll show whether or not the button is enabled for this
question.
Updated comments + formatting.

sender.ino: Fixed a bug in the BUTTON_EVENT message handler causing the
team number to not get set.
Commented out reading in the ButtonEvent data in the BUTTON_EVENT
handler.
Updated comments + formatting.